Added a CardView which you can get and clone without taking ownership
1. Card doesn't implement Clone or Copy
2. Card has a private constructor

These things together mean we know we won't run into a scenario where we
have cloned a card and played/discarded it twice at compile time. But in
order to get information about someone else's hand we'd like to return a
Vec of their cards. To do this, we create a new type CardView that can
be created from a Card. But since it's a new type it can't mistakenly be
appended to the discard pile or to a player's hand.